Move charts from ReportFragment to RevenueExpensesFragment
