+++ app/src/main/res/layout/fragment_report.xml
@@ -25,60 +25,4 @@
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@id/tab_layout" />
 
-    <com.google.android.material.textview.MaterialTextView
-        android:id="@+id/textview_start_month"
-        style="@style/Widget.MaterialComponents.TextInputLayout.OutlinedBox"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:layout_marginHorizontal="15dp"
-        android:layout_marginTop="30dp"
-        android:background="@drawable/background_textview_round_outline"
-        android:drawablePadding="10dp"
-        android:gravity="center_vertical"
-        android:hint="@string/report_start_month"
-        android:padding="10dp"
-        app:drawableStartCompat="@drawable/ic_baseline_calendar_month_24"
-        app:layout_constraintEnd_toStartOf="@+id/textview_end_month"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <com.google.android.material.textview.MaterialTextView
-        android:id="@+id/textview_end_month"
-        style="@style/Widget.MaterialComponents.TextInputLayout.OutlinedBox"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:layout_marginHorizontal="15dp"
-        android:layout_marginTop="30dp"
-        android:background="@drawable/background_textview_round_outline"
-        android:drawablePadding="10dp"
-        android:gravity="center_vertical"
-        android:hint="@string/report_end_month"
-        android:padding="10dp"
-        app:drawableStartCompat="@drawable/ic_baseline_calendar_month_24"
-        app:layout_constraintBaseline_toBaselineOf="@+id/textview_start_month"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toEndOf="@id/textview_start_month" />
-
-
-    <com.github.mikephil.charting.charts.BarChart
-        android:id="@+id/barChart"
-        android:layout_width="match_parent"
-        android:layout_height="0dp"
-        android:layout_marginTop="30dp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="@+id/textview_start_month" />
-
-    <com.github.mikephil.charting.charts.LineChart
-        android:id="@+id/lineChart"
-        android:layout_width="match_parent"
-        android:layout_height="0dp"
-        android:layout_marginTop="30dp"
-        android:layout_marginBottom="80dp"
-        android:visibility="gone"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/textview_start_month" />
 </androidx.constraintlayout.widget.ConstraintLayout>
